backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Pillow==10.0.0
python-dotenv==1.0.0
bcrypt==4.0.1
numpy==1.25.2
opencv-python==4.8.0
pymongo==4.5.0
python-magic==0.4.27
Werkzeug==2.3.7
gunicorn==21.2.0 